abstract coco
